requirements.txt
asgiref==3.8.1
certifi==2024.7.4
contourpy==1.3.1
crispy-bootstrap5==2024.2
crispy-tailwind==1.0.3
cycler==0.12.1
DateTime==5.5
distlib==0.3.8
dj-database-url==2.3.0
Django==5.0.6
django-crispy-forms==2.3
django-environ==0.11.2
django-phone-field==1.8.1
django_templates==0.0.13
extends==0.3.0
filelock==3.15.4
fonttools==4.55.3
gunicorn==23.0.0
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
psycopg2==2.9.9
psycopg2-binary==2.9.10
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
tabulate==0.9.0
tk==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
win10toast==0.9
zope.interface==7.1.1
